+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Fix error running OCR from dialog.
   Closes #59 (OCR Fails: `int` object has no attribute `get_size`)
+* Eliminate extra tesseract call when saving PDFs.
 
 
 ## 3.0.3 (2026-04-03)