+++ changelog.md
@@ -3,7 +3,7 @@
 * Fix error running OCR from dialog.
   Closes #59 (OCR Fails: `int` object has no attribute `get_size`)
 * Eliminate extra tesseract call when saving PDFs.
-* Fix message “Some pages have not been saved" when all pages have been saved
+* Fix message 'Some pages have not been saved' when all pages have been saved
   Closes #61