+++ changelog.md
@@ -1,4 +1,4 @@
-## 3.0.4 (2026-04-10)
+## 3.0.4 (2026-04-11)
 
 * Fix error running OCR from dialog.
   Closes #59 (OCR Fails: `int` object has no attribute `get_size`)